+++ tests/unit/rock/test_build.py
@@ -20,7 +20,7 @@ import itertools
 import craft_platforms
 import pytest
 import pytest_check
-from craft_platforms import rock
+from craft_platforms import _errors, rock
 
 SAMPLE_UBUNTU_VERSIONS = ("16.04", "18.04", "20.04", "22.04", "24.04", "24.10", "devel")
 
@@ -42,6 +42,11 @@ SAMPLE_UBUNTU_VERSIONS = ("16.04", "18.04", "20.04", "22.04", "24.04", "24.10", 
             )
             for version in SAMPLE_UBUNTU_VERSIONS
         ],
+        # Legacy base strings containing colons
+        ("ubuntu:20.04", None, craft_platforms.DistroBase.from_str("ubuntu@20.04")),
+        ("ubuntu:22.04", None, craft_platforms.DistroBase.from_str("ubuntu@22.04")),
+        ("bare", "ubuntu:20.04", craft_platforms.DistroBase.from_str("ubuntu@20.04")),
+        ("bare", "ubuntu:22.04", craft_platforms.DistroBase.from_str("ubuntu@22.04")),
     ],
 )
 @pytest.mark.parametrize(
@@ -71,6 +76,19 @@ SAMPLE_UBUNTU_VERSIONS = ("16.04", "18.04", "20.04", "22.04", "24.04", "24.10", 
         *[
             pytest.param(
                 {
+                    architecture.value: {
+                        "build-on": architecture.value,
+                        "build-for": architecture.value,
+                    },
+                },
+                {architecture.value: [(architecture.value, architecture.value)]},
+                id=f"explicit-scalar-{architecture.value}",
+            )
+            for architecture in craft_platforms.DebianArchitecture
+        ],
+        *[
+            pytest.param(
+                {
                     "my-platform": {
                         "build-on": [
                             arch.value for arch in craft_platforms.DebianArchitecture
@@ -88,9 +106,30 @@ SAMPLE_UBUNTU_VERSIONS = ("16.04", "18.04", "20.04", "22.04", "24.04", "24.10", 
             )
             for build_for_arch in craft_platforms.DebianArchitecture
         ],
+        *[
+            pytest.param(
+                {
+                    "my-platform": {
+                        "build-on": [
+                            arch.value for arch in craft_platforms.DebianArchitecture
+                        ],
+                        "build-for": [build_for_arch.value],
+                    },
+                },
+                {
+                    "my-platform": [
+                        (arch.value, build_for_arch.value)
+                        for arch in craft_platforms.DebianArchitecture
+                    ],
+                },
+                id=f"build-on-any-for-scalar-{build_for_arch.value}",
+            )
+            for build_for_arch in craft_platforms.DebianArchitecture
+        ],
     ],
 )
 def test_build_plans_success(
+    check,
     base,
     build_base,
     expected_base,
@@ -105,9 +144,9 @@ def test_build_plans_success(
     )
 
     for build_item in build_plan:
-        with pytest_check.check():
+        with check():
             assert build_item.build_base == expected_base
-        with pytest_check.check():
+        with check():
             assert (build_item.build_on, build_item.build_for) in platform_archs[
                 build_item.platform
             ]
@@ -264,3 +303,9 @@ def test_build_plans_bad_base(base, error_msg):
 def test_build_plans_bad_architecture(platforms, error_msg):
     with pytest.raises(ValueError, match=error_msg):
         rock.get_rock_build_plan("ubuntu@24.04", platforms)
+
+
+def test_bare_base_no_build_base() -> None:
+    """Make sure that an error is raised if base=="bare" but build-base==None"""
+    with pytest.raises(_errors.NeedBuildBaseError):
+        rock.get_rock_build_plan("bare", platforms={"amd64": None}, build_base=None)